pollcompat.c: improve device handling
When building without POLLCOMPAT_ALWAYS defined, the system poll call
will be used in the limited circumstances in which it will actually
function correctly.  This is the default build configuration.

Improve the logic so that not just isatty(...) devices cause the
system poll() function to be avoided, but any problematic devices
(as the logic should have worked in the first place).

Additionally improve handing of bits set in the errorfds set returned
by select.  If a bit set in errorfds is ALSO set in the readfds set
then no longer set POLLERR.  Some devices (e.g. /dev/null and /dev/zero)
can cause this spurious condition to appear.  The next read will
return without blocking in any case and if there's a real error it
will be reported at that time.  This tremendously improves pollcompat
compatibility when working with such devices (/dev/null being common).