pollcompat.c: POLLOUT parity
The select() "feature" where some devices cause the errorfds bit to
always be set alongside the readfds bit also affects writefds.

In other words, if the same bit is set in both writefds and errorfds
there may not actually be any error.  The /dev/null device is one
example of a device that produces this condition.

Apply the same "kludge" for POLLOUT that POLLIN already enjoys --
do NOT set POLLERR when an errorfds bit is set unless NEITHER the
corresponding readfds nor the writefds bit is set.