pollcompat.c: workaround unreliably cleared select errorfds bits
Setting a fd's errorfds bit without also setting at least one of its
two other select bits (readfds and/or writefds) can result in the
failure of the errorfds bit to actually be cleared on return even
when there's no actual error on the file descriptor.

This can cause nasty problems.

Since we already suppress returning a POLLERR when both an errorfds
bit and either of the other bits come back set, only set the incoming
errorfds bit when either POLLIN and/or POLLOUT is requested.

In the unlikely event that there exists some error condition that will
only be checked for when the errorfds bit is set, it should be caught.
And in the further unlikely case where the result bits have only the
file descriptor's errorfds bit set, a POLLERR will actually be returned
by this code.  But that's now pretty much the only way it can happen.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>