pollcompat.c: improve EPIPE handling
Ideally, POLLERR (or even POLLNVAL) ought to be set when trying to
POLLOUT on a widowed pipe (if SIGPIPE has not already terminated
the process).

Unfortunately, the only way to tell is to try to write to the file
descriptor.  That's bad and we can't do that.

BUT, it turns out that the built-in system poll does actually work
on normal, unnamed (i.e. NON-mkfifo) pipes and will set POLLNVAL
in that case.

However, both named (i.e. mkfifo) and unnamed pipes return true for
the S_ISFIFO check.  BUT, if we have the F_GETPATH fcntl we can
tell the difference as it will succeed on mkfifo pipes (even if the
file system node has been unlinked since it was opened) but fail
on pipe(2) created pipes.

Add this kludge test so that the system poll can therefore be used
to at least get an error condition for widowed unnamed pipes when
using pollcompat with compatible fds.  And that is a very good
argument for not always defining POLLCOMPAT_ALWAYS when building.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>